Support impl args buffer in CMImplParam (XeHP-)
CMImplParam was extended to support Local Size, Group Count, Printf
Buffer Ptr implicit args in external and indirect functions for
pre-XeHP_SDV platforms.

Implicit args buffer predefined variable was utilized to implement the
feature: the variable is initialized in the kernel prologue with the
value from the corresponding implicit argument, later when different
implicit args are requested in external and indirect functions, the
buffer pointer is retrieved from the variable and the required args are
loaded from the buffer.