Reland "Fix the calculation of jitinfo's usesBarrier in finalizer."
Calculate the number of barriers when translating the vISA barrier
instructions. For named barriers, derive the number from barrier ID
instead of the number of send barrier instructions. For the legacy
barrier, set it to 1 when there's a barrier instruction.

Also when propagating the barrier information from callee to caller,
change the caller's usesBarrier to callee's value if callee's value is
bigger than caller's.