Fix dangling arguments that leads to random runtime results with stack calls
In the case of DeSSA where we have phi nodes for function arguments, we
do not create moves for those phi nodes. In union-find data structure,
we need to assign the leaders of arguments (phi nodes) to the symbol
table in order to get correct reference; otherwise, it will create a new
dangling CVaraible.