Use binary linking when invoke_simd target definition is not present
In previous implementation we always assumed, that definition for invoke_simd target will be provided via .visaasm coming from VC compilation. This meant that direct call was always used and LTO would be triggered.
This change introduces a possibility to use standard binary linking if the .visasm definiton is not provided, to cover ISPC use-cases.