Fix a weird 'bitcast' from i64 to i64.
Expaned values are directly reused through 'bitcast' if the source is an
i64 value. But, if that 'bitcast' is a one from i64 to i64, we miss the
chance to reuse the already expanded values as it's assumed that, if the
destination is an i64 value, the source won't be an i64 value. The
results in incorrect emulation code.