Fixed the alignment for indirect float/half dst
Float/half operations requires all operands to have the same offsets.
If it is indirect, we have to assume they are not aligned so that
raw mov insts are generated. Raw mov instructions can be changed to
int type, as int indirect addressing is more flexible.

This change make sure that fixDstAlignment() will be invoked for
indirect float/half dst.