[Autobackout][FuncReg]Revert of change: 5277e60786a6e4bf311c22faf99d79deb9e56e66  Change the behavior of FP64 kernels poisoning
PoisonFP64Kernels pass removes all functions with double type usage. It
removes regular functions as well as kernel functions. Removing kernel
functions cause an issue with multitarget AOT compilation, because
SYCL FE calls clCreateKernel/zeKernelCreate API calls for all targets,
not only the one that it is being run on. It leads to reporting an error
by L0/OpenCL Runtime even if an application is being run on a platform
with native fp64 support.