Mark stores as volatile so they're not optimized away.
Stores emitted in kernel to write local ids to frame are never
referenced in a load instruction in the kernel. This causes llvm
optimizer to treat the stores as dead and they're subsequently optimized
away. This change marks those stores as volatile to indicate that some
callee may load data written by these stores and they shouldn't be
optimized away.