Always compute EMask for BBs if inlineAsm is present
The active execution mask is cached for each BasicBlock to prevent creating duplicate emasks.
However, if inline asm calls are present, they may introduce "labels" that function as new BBs
that LLVM does not track. In the case of divergent control flow, the "else" branch may reference
an uninitialized EMask predicate.
This change forces computing the EMask for each BB if inline asm is present in the module.