Fix problem with cached paths for llvm prebuild
Force setting correct paths to source for lld and clang when we build
llvm prebuilds. Previously the paths can be cached before with the wrong
variables. Now, as we know that external/llvm/prbuilds/CMakeLists.txt is
the first Cmake file and variable ${IGC_OPTION__LLVM_SOURCES_DIR} must
be set. We can safely force writing the new value to the paths.