Fixed incorrect types for rotate inst
As rotate needs to keep its dst/src0 type the same and no widening or
narrowing should be done on those operands, make sure copyProg and defHoist
do not change the type of dst/src0.

Also, make sure to use the larger type of dst/src0 as rotate type to be conservative.
This is to handle the case in which imm src0's type has been changed.