ZEBinary: Enable forceGlobalMemoryAllocation when there is private global variables
A global variable is assumed to have global or constant address space and should be allocated on global/stateless
memory. While it is possible that a global variable given from the input has private address space (or no address space)
which confuses IGC and results in inconsistent decision when deciding its pointer size.
E.g. In CodeGenContext::getRegisterPointerSizeInBits, the pointer size of such variable will be 32 (according to private
address space) assuming it is allocated on scratch, but in other passes it is allocated on global memory since it is a
global variable.

This is a WA fix to force all private variables be allocated on global to avoid the confusion.
The WA is applied to ZEBinary path only. Patch-token path has another WA to force constant address space to those
variables at ProgramScopeConstantAnalysis. It's not applicable to ZEBinary since ZEBinary relies solely on symbol/relocation
for global variable access, but patch-token relies on implicit argument "global/constant base".