Library compilations cannot use scratch space for private memory
SS Private memory is disabled for when stackcalls are enabled, including indirect calls.
However when a module is compiled as a library, i.e. there is no kernel function, only external-linkage library functions, we fail to detect that module as having a stackcall.
In this case we need to disable SS private memory as well since the dynamically linked callers will also have it disabled.