Fix crash when ShaderDumpEnableAll is enabled
`PassManager::add(P);` may call `delete P;` when it detects that analysis
has been already processed and it doesn't need to be added to the pass
manager. Then at the very beggining of `addPrintPass` function, IGC
dereferences `P` which points to a deleted object, causing a crash.

This fix takes the logic responsible for verifying whether current pass
is an analysis pass out of `addPrintPass` function to avoid problematic
pointer dereferencing.