Change vISA's -fusedCallWA from bool to uint.
Change vISA's option -fusedCallWA from bool to uint, so it can be used
to control software wa without hw fix as well as with the partial hw fix.

This submit just changes the option's type, and there is no functional change.
The WA with the partial hw fix will be added later.