Lower LLVM intrinsics for integer min, max, abs
With the switch to LLVM 12+ (LLVM 14 in particular), new LLVM
intrinsics get generated for the known test cases. Lower these
intrinsics to GenX ones in accordance with the following map:
_________________________________________________

| llvm.abs(i32, i1)   |  llvm.genx.absi(i32)     |
_________________________________________________

| llvm.smin(i32, i32) | llvm.genx.smin(i32, i32) |
_________________________________________________

| llvm.smax(i32, i32) | llvm.genx.smax(i32, i32) |
_________________________________________________

| llvm.umin(i32, i32) | llvm.genx.umin(i32, i32) |
_________________________________________________

| llvm.umax(i32, i32) | llvm.genx.umax(i32, i32) |
_________________________________________________

`i32` is chosen as a mere example - similarly to FP-typed intrinsics,
these functions operate on any integer/integer vector types.