Refactor call wa
Refactor the call wa to prepare for supporting diverged fused call in general.

As diverged fused call isn't supported by hardware and the software cannot
guarantee the uniform call targets of fused call, the software needs to
conservatively duplicate calls for bigEU and smallEU.  The detail will be
added later.

No Functional change for this submit.