call ws for partial hw fix
fused call has a partial HW fix, which needs to remove predicate call. Also
any indirect call might have different targets for bigEu and SmallEu,
therefore, needs to have two calls, one for bigEu and the other for smallEU.

The new code is off now and will be turned it on in next submit.