Fix for stack function map update, when vAsmTextBuilder was used.
If CISABuillder is used in text mode  (e.g. when ShaderOverride is turned on) we need to update the map of stack functions after parsing the .visaasm.
In case of ESIMD and SPMD interop we parse additional visaasm passed in the input. In this scenario the stack function map was not updated correctly.