ESIMD+SPMD interop: parse additional vISA asm passed in the input
For ESIMD and SPMD interoperability we need to support a use-case, when SPMD code is called from ESIMD code (so called "invoke_sycl" use-case).
For this use-case we want direct calls and vISA linking. The SPMD function visaasm is passed in the input arguments.
Same approach is used for "invoke_simd" use case on SPMD path.