Skip verifying live var w/o physical reg assigned.
Skip verifying a live var that has no physical reg assigned. Currently there
might be some cases of a variable not actually used but marked as live-in and
live-out. For example, spillHeader var is created for save/restore spill/fill
code in stack call case, and is not used when expanding spill/fill using LSC.