Fix value tracker algorithm
The issue is that ValueTracker marks all instructions it was processing as
visited, but it may happen that some instructions need to be revisited
to properly track an image. Example:

```
   alloca    image (kernel argument)
      |        /
     GEP      /
    /   \    /
 load   store
 |
image_write
```

Tracking starts at `image_write` instruction. It goes through `load` and
`GEP`, marking them as visited. Since `alloca` keeps the image, the
algorithm needs to go down through the tree now. To reach `store`
instruction, it needs to go through `GEP`, but it can't, since it
is marked as visited. In, short, marking instructions as visited when
going up the tree leads to cutting unvisited tree branches.

This change removes marking instructions as visited when going up the
tree.