Fixing bad calling sequence
Lowering intrinsincs resulted in bad calling sequence. If demanded
intrinsic call is found in basic block, it has to be moved at the
beginning of it to make sure that it can be used safely.

Little optimization added: if there is already simdLaneId or simdSize
call in current BB - don't call it again and move it to the beginning of
function.