[Autobackout][FuncReg]Revert of change: 76f1ae0c10dba29246ce0a8afb6a39d56b3ecd21  [IGC vISA] Fix the src2 reg region issue.
For src2, Register Regioning patterns where register data bit location
of the LSB of the channels are changed between source and destination are
not supported except for broadcast of a scalar. The src2 regioning should
have been conformed in fixMADInst() or fix3SrcInst() which are before
fixByteXBarRestriction(). But to resolve the byte XBar restriction, the dst
region has been changed. So we need to fix src2 region accordingly.