Add CROSS_THREAD_OFF_R0_RELOCATION runtime symbol in ZeBinary builder if it is used in the program
CROSS_THREAD_OFF_R0_RELOCATION symbol was added in ZeBinaryBuilder only if EnableGlobalStateBuffer registry flag was set.
This relocation in the kernel is set when vISA_emitCrossThreadOffR0Reloc option is set. VC backend adds this option without checking for EnableGlobalStateBuffer registry setting for subtargets that have thread payload in memory.

This change makez ZeBinary builder add the runtime symbol if the relocation is detected in the program.