Reland ZEBinary: support "-disable-zebin" option and remove ZEBin to CTNI fallback path
Add IGC part of supporting option "-disable-zebin" for disabling ZEBinary.
This is for the preparation of enable ZEBinary by default.

The change also remove the fallback path of ZEBinary to CTNI/patch-token.
There is codegen difference between ZEBinary and CTNI, but not just the ouput formats.
The fallback path is no longer applicable.
By removing the fallback path, the additional option, CompOptions::EnableZEBinary, is removed.

Now the IGC regkey "EnableZEBinary" is the single/unified option for enable/disable ZEBinary
across IGC.