Reland ZEBIN: Add NT_INTELGT_ZEBIN_VERSION for whole ELF versioning.
The description of the new ELF note NT_INTELGT_ZEBIN_VERSION represents
the ZEBIN ELF file version that reflects the attribute and section
changes. The content is stored in a nul-terminated string and the format
is "<Major number>.<Minor number>". For example, it's "1.16\0" with this
change.

Currently the version information is in both .ze_info and the new ELF
note in order not to break any existing tools. After switching to the
ELF note, we can decide if we want to remove the version in .ze_info or
keep the redundant info in both places.

Also separate version information as a version.md from zeinfo md file
because now it should reflect all ELF section and zeinfo attribute
changes.