[Autobackout][FuncReg]Revert of change: 4096684a19f67ef83064f9a19e1c7c6e411b1f00  VisaLinkerDriver: support for invoke_sycl scenario in ProcessElfInput
 L0 experimental extension "ZE_experimental_module_program" adds the ability to pass multiple SPIR-V modules for compilation. This extension will be used by ISPC for visa-linking scenarios.

This change alters the processing of Elf in IGC. The ProcessElfInput function was used for module linking for API calls such as clLinkProgram. Only SPMD modules were present in the Elf, so LLVM linking could be used to merge the modules. Now this function needs to check for the "flavor" of each SPIR-V passed. If any ESIMD module is present, VIsa linking needs to be done.