IGA: fixed swsb setter for special ARF dependency of potential shoot-down instructions
Extra sync.nop for preserving possible-shoot-down instruction's sbid caused violating the rule:
"When certain Arch Registers (sr, cr, ce) are used, the very next instruction requires dependency
to be set on all pipes {A@1}".

Create another sync to move A@1 to before the above mentioned sync.nop.
We don't track if the A@1 is inserted for this rule or other dependencies. Conservatively
move A@1 always when seen.