Remove unsafe pattern match optimization for fpext
We are incorrectly removing the following pattern...

  %a = fptrunc float %0 to half
  %r = fpext half %a to float
=>
  ret float %0

This is not safe as proven by alive tool for cases that exceed HALF_MAX and HALF_MIN (really large values and really small values) which will turn into +/- infinity and then be casted back as +/- infinity. Also float values when converted to half with rounding done and if in the middle their is no telling of which way to cast back.

Note that this may work sometimes but depends ultimately on the floation point value.