Add -stopafter vISA option to stop compilation after a pass
Add a -stopafter <pass> vISA option to stop compilation after the pass. It is
intended for vISA offline debugging and writing LIT-style tests. When the pass is
executed, it emits the result IR into a file and early exits the compilation.
The option has no effect if no pass with the same name is executed.