Add VISA_DEBUG() macro to conditionally emit debug information
Add a VISA_DEBUG() macro to conditionally emit debug information, similar to
how LLVM_DEBUG() works. The macro is disabled by default, the "-debug-only <list>"
vISA option enables it. If <list> is equal to "all", the macro is enabled
globally. Otherwise, the macros is conditionally enabled in the passes named
in the comma-separated list. There is only one current debug pass at a time,
which can be set internally via setCurrentDebugPass(). In addition to the
passes listed in the optimizer, other valid pass names include "translate"
(for vISA to G4 translation), "CFG", and "encode".

The macro is enabled only for offline vISA compilation and when NDEBUG is not set,
so it does not affect compile time.