Allocate second RTGlobal structure for sync DispatchRays
Sync DispatchRays does not need the async part of the ray stack.
We can use RTGlobals struct that has RT mem base ptr set to the bottom of async stack,
allowing sync dispatch to use the whole raystack as the sync stack.
We allocate this modified struct in the indirect buffer, right after the original RT globals struct.
We can't discard the original because it might still see some usage.