Separate code path not relevant to current platforms
Split send is supported from SKL+. So height of message declare can be
set to 1 for SKL+. Only when split send is not supported, ie BDW, do we
need to use old code path. This change adds a condition to check whether
split send is supported, and if so, use height = 1.

Added a condition for early return when stack calls are used. This
condition returns reserved temp register as message header.