Allow boolean flag to have a value
Before this change,  visa flag is in the form:
    -flagName
After this change, it will be:
    -flagName [0|1]
When flag's value is present, it will set the boolean flag based on the
value (0 for false, 1 (or not equal to zero) for true).  If the value is
not present, the flag behaves the same as before (inverse of its default).