(NFCI) Use getNumSrc() instead of G4_MAX_SRCS to iterate an instruction's sources
A lot of legacy vISA code iterate through an instruction's sources using
G4_MAX_SRCS as the loop upper bound. Replace them with getNumSrc() when
possible, as the vast majority of instructions have 1-3 sources but
G4_MAX_SRCS is 5 due to some rarely used instructions.

The goal is to eventually avoid null checks when iterating an instruction's
sources, but leave that as a future commit as it might cause crashes on
some obscure instructions.