Fix condition to convert current iteration in to fail safe
To convert current iteration in to fail safe, we need to ensure that
BuiltInR0 is available everywhere in the program. Unless BuiltInR0 is
marked live-out, RA is free to assign r0 to some virtual variable after
last use. If r0 value is lost then we cannot emit HWord scratch message
for spill/fill.