Invoke_simd: accept SPMD+ESIMD SPIR-V modules in linking API
Applications that use invoke_simd functionality use different ways to enter IGC:
1. We may get a single ESIMD+SPMD SPIR-V module
2. zeModuleCreate with extended descriptor might be used - this way many SPIR-V modules may be passed, combined in an Elf file by Neo.
3. linking API might be used by DPCPP runtime. This eventually ends up in the same path as 2.

In paths 2. and 3. mentioned above we have a list of SPIR-V modules to link. Until now, we assumed that only SPMD or ESIMD SPIR-V is accepted there.

This change adds the ability to process SPMD+ESIMD modules there as well.