g4 dumps uses igc dump prefix.
Use the same igc prefix for g4 dump by default so that it is clear
which g4 is for which igc shader.

By default, those g4 dumps will be in the same directory as igc dumps
unless vISA's flag -dumpToCurrentDir is set. If -dumpToCurrentDir is
set,  g4 dumps will be in the current directory.