Traverse to base G4_RegVar when computing spill offset
When an address taken variable is spilled, a temporary is created for
filling original value. This temporary needs a spill offset slot to load
data from. The temporary should load from spill slot of
original variable. When computing spill slot to use for the temporary, therefore, we should
traverse to base G4_RegVar.