(NFCI) Refactor functions that check if an operand is an address or indirect, second try
Fix a bug in the previous commit where we did not consider that a direct
address operand may be a destination as well. It causes the dependency
between an address definition and indirect operand to be missed, and local
def-hoisting generates wrong code as a result.

Original description:
The existing G4_Operand->isA0/isAddress functions return true for both direct
address operand (a0.0<1>:uw) and indirect register operand (r[a0.0]<1;1,0>:ud).
Such ambiguity affects readability and is likely error-prone (e.g., caller
may only expect address operands and not indirect operands). Refactor it
so that isDirectA0/isDirectAddress only returns true for address operands.
RegRegion->isIndirect() should be used instead to check for indirect operands.

While the intent of the commit is to not introduce functional changes, some
existing uses of isA0/isAddress are already ambiguous in that the context of
the call site suggests it is expecting either address or indirect operand,
but the old code will return both. For such cases I decided to go by my
perceived intent of the caller and choose either address or indirect operand.
This might result in regressions, but I think it's better to expose such bugs
than hide them.