ZEBIN: Put a const string literal into .data.const.string section only if used by printf only.
Compiler can not distinguish a char array between a string literal. So,
to be safe we can only place a const string literal into the read-only
str section that is accessed by CPU when the string literal is used only
by printf. For all other cases, we need to place it into the read-only
data section.