Add condition to determine if spill is partial write
Scratch msg is used for spill/fill when stack ABI is not used. For pre-PVC, scratch msg honors EM when elem size == 4. For PVC, EM is treated as being SIMT1. This change adds condition to determine whether spill msg is to be treated as partial write.

Bug shows up when LSC is forced for non-stack call spill on DG2.

Bug also shows up when HW scratch msg is used by default but offset
exceeds encoding limit and we use LSC or OW msg.