ZEBIN: Fix printf for platforms without A64 support
String literal pointer is a pointer to CPU memory allocated through
malloc on Neo side, so we shouldn't use llvm datalayout to decide what
is the size of pointer. Since IGC doesn't know what is the CPU pointer
size, we always treat this as 8-byte to allow passing pointer for
both 32 and 64 bit applications.