[Autobackout][FuncReg]Revert of change: 8f1feecf039c0549fadb66989bc0a000ca83a423  invoke_simd: getting list of exported and imported functions from SPIR-V modules
When invoke_simd builtin is used we may use binary linking or vISA linking. In binary linking case indirect calls are used, in vISA linking we need to force direct calls.

The definitions of the direct-call functions may be present in modules that were not yet compiled. This change adds an input field with directly called function names to scalar path in IGC. The list is based on exported and imported functions in the input SPIR-V modules: if a function is imported in ESIMD and exported in SPMD, or vice-versa, it is added to the list of direct-call functions for visa linking.