Use shorter temporary file names for ZEBin debug info
Make tempfile names independent from the kernel name so that the name's length
is guaranteed to be acceptable for the linker. Store the names in a JSON to
preserve their mapping onto kernel names.

The new infrastructure is also re-used for VectorCompiler, which is made
possible through additional refactoring of `static` helpers into
`CGen8CMProgram` class methods.

Co-authored-by: Stanley Gambarin <stanley.gambarin@intel.com>