Deprecate R0 function argument, and always use GenISA_getR0 instead
R0 as an implicit argument is statically defined as 8 x DWORD. However the size should be platform dependent.
Since a dedicated register is always used to store R0, and we already support the intrinsic GenISA_getR0 for
stackcalls, which maps to this physical register, we can deprecate the function argument and always use the
intrinsic for R0 access in the shader.