ZEBIN: Recognize oneapi printf as opencl printf.
According to OPENCL C printf spec, the use of the conversion specifier
"s" is very restrictive. A printf wrapper function could cause the
format string and the printf arguments of a printf call are from
function arguments. This leads to increased complexity to do printf
analysis and resolution in OPENCL, and also violates the spec actually.
Here's the printf wrapper example found in DPC++.
https://intel.github.io/llvm-docs/doxygen/ext_2oneapi_2experimental_2builtins_8hpp_source.html

All printf wrapper cases should be treated as errors, and will need to
enhance the compiler error reporting. However, to support ONEAPI printf,
this change treats ONEAPI printf calls as OPENCL printf calls.

This change also improve the checks for in-between uses from a GV to
a printf call.