ScalarArgAsPointerAnalysis pass - support for variable array index
Kernel can take as input a struct with array of pointers. If pass can't deduce which
pointer is used for access to global memory (when array index is variable with
unknown value), mark all pointers (all array indices) as used for access. This could
give false positive results, but it is accepted as a better outcome than not matching
any pointer at all.