Support dynamic resolution for generic addresses
Load/store instructions working with generic memory must be resolved to loads/stores
which use either local or global memory before GenXLoadStoreLowering pass.

For genx64 the local pointer is 64bit but only 32 of them will be used as actual address,
so we use other 32bits to store a tag. Then the load/store to/from generic memory
will be resolved based on this tag.