(NFC) move local RA's physical register usage summary to be inside the RA pass
The local RA pass creates a bitset of physical registers used by each basic
block for later use by the hybrid RA algorithm. This PhyRegSumnmary mapping
for each BB is stored in the FlowGraph class for some reason. Move it to the
GlobalRA class where it more naturally belongs, and this also allows them to
be freed at the end of RA. Also clean up some legacy debug verbose macros.