(NFCI) Avoid passing each basic block's scheduling stats in FINALIZER_INFO, second try.
vISA FINALIZER_INFO struct has each basic block's scheduling stats (weighted and
unweighted stall cycles), which is used by IGC both for stat reporting and also
in SIMD size selection heuristics. This requires a dynamic allocated array to
store the information. Avoid this by having vISA sum up the stats for the shader
instead. Also move the basic block count information in FINALIZER_INFO out of
the post-RA scheduler pass as it is not necessarily enabled always.

The first submission was backed out due to test regressions, but it seems the
test is unstable and has sporadic failures.