(NFC) Avoid redundant string copy for variable names, and rename createDeclarNoLookup to createDeclare
When shader dump is enabled, vISA is unnecessarily creating two copies of the
variable names. Remove the duplicate copy to lower memory footprint. A string
copy is still generated when shader dump is not enabled as the passed in
char* may be stack allocated, but for release builds IGC uses empty string
for variable names anyway so it should not incur overhead.
"
Also rename IR_Builder::createDeclareNoLookup to createDeclare, as "NoLookup"
doesn't make sense anymore; the function always returns a new G4_Declare.