[Autobackout][FuncReg]Revert of change: 50c42701747460e1adb9bce6544f3d8a7d79b579  Mark LSC prefetch builtin as `optnone` to ensure immediate offset
When building the BiFModule, Clang 13+ can break LSC prefetch builtin semantics
at the IR level. `SimplifyCFG` looks to reduce the number of call instructions
with the same set of argument values, and for the current implementation of the
builtin, we end up getting PHI values for the offset argument, whereas an
"immediate" integer value is expected.

As an alternative, we could allow the optimizations and then "undo" the CFG
changes in `LSCFuncsResolution` or a separate IGC pass. However, this approach
feels more like a workaround for our inability to enforce the immediate
argument at the level of builtin source code/Clang build command. Unless
there's any noticeable performance impact, marking LSC prefetch as `optnone`
should be the more precise solution for IGC/LLVM 14 compatibility.