Mark LSC prefetch builtin as `optnone` for immediate offset (take 2)
When building the BiFModule, Clang 13+ can break LSC prefetch builtin semantics
at the IR level. `SimplifyCFG` looks to reduce the number of call instructions
with the same set of argument values, and for the current implementation of the
builtin, we end up getting PHI values for the offset argument, whereas an
"immediate" integer value is expected.

As opposed to the original approach in commit 50c4270, do not remove `optnone`
and `noinline` as part of `ProcessFuncAttributes`. `LSCFuncsResolution` will
replace all of the SPIR-V builtin calls with GenISA intrinsic calls anyway;
meanwhile, the attribute removal leads to other "argument immediacy-breaking"
optimizations early in IGC pipeline.