(NFC) Remove unnecessary include of VISADefines.h
VISADefines.h is #included by a number of files even though it has a small
number of #defines, most of which unused. Minimize the dependency on this
header file by removing unnecessary dll_export macros on functions that
don't need to be exported. Also remove __thread qualifier as they are
used by global variables for shader dump and timer, and we don't care if
these are thread safe.