Fix VISA index in cross and per thread prolog
Some instructions in cross and per thread prolog were being constructed
and added to IR Builder's inst list. G4 instructions created in this
manner take max VISA index as their own VISA index. This is useful when
creating IR during initial phase as VISA index increments appropriately.
However, this is a problem when creating new instructions to insert
somewhere in existing code as new instructions inherit potentially OOB
VISA index. When VISA index is OOB, it impacts debug info as IGC treats
such OOB VISA index as end of the compilation unit.

The fix in this case is to invalidate VISA index of instructions created
for cross and per thread prolog as they are not lowered from a
particular VISA index.