Replace fcmp_one to fcmp_une condition for minmax instruction
During the minmax comparison, we should check if any input value is NaN
and return the other one value (non-NaN). If both are NaN, return NaN.
For this we need to use fcmp_une condition instead fcmp_one.
Unordered comparison gives true if any of operands (or both) is NaN.
Unlike une, the ordered comparison return true
if neither number of operands is NaN.