(NFC) Always use root variable name when printing operand in G4_IR
When an aliased variable appears in an operand in G4_IR, always print the
root variable's name, plus the byte offset if it's non-zero. For example, if
we have

.declare V0  rf=r size=8 typed=ud
.declare V1  rf=r size=4 type=ud alias=V0+4

an use of V1 will be printed as V0_4(0,0)<0;1;0>:ud
It makes it much easier to read a variable's def/use.