[Autobackout][FuncReg]Revert of change: 730461e512eb0e07885f0bbce39732278b1be683  Allows per-function retry mechanism for stackcall programs
Building on retry for stackcalls, this change also enables/disables individual passes on the second try such that FunctionPass level optimizations are ran on a per-function basis.
This should help with performance on retry, where we don't use the optimization states of the second try on functions that did not spill.