Patch LLVM 14 LICM pass
Remove too strict restrictions in LICM pass.

The reason for removal of above restriction  was that it took a very strict
approach to the Convergent attribute, which caused missed optimization
opportunities in cases where it was safe to do so.
The decision is based on the discussion in LLVM RFC
https://reviews.llvm.org/D90361?id=303195
This patch should be considered obsolete if LICM introduces a more
advanced approach to the Convergent attribute in the future version of
LLVM