[Autobackout][FuncReg]Revert of change: 8c44636b29867c28838ddd652224ae6d08b7635d  Ensure rematerialized scalar doesn't get re-rematerialized.
When inserting spill/fill code, we rematerialize scalar imm mov. Such
movs create a new temp to load scalar imm value. This temp is not an
infinite spill cost variable as its use may span multiple instructions
in case of spatial reuse.

Since the temp is not an infinite spill cost variable, it's possible for
RA to re-spill the temp in subsequent iteration. This causes spill/fill
module to re-rematerialize the scalar imm mov, which could lead to an
infinite cycle.

This change fixes the problem by storing list of temp vars created in
earlier RA iterations. This way we avoid re-rematerialization problem.