Always enable LICM on retry when there are stackcalls
Default behavior disables LICM on the retry compilation.
Since LICM is a LLVM pass, we cannot enable/disable it at a per-function level.
Based on performance testing, we found that enabling LICM when stackcalls spill gave better performance than with it disabled, so this change always enables LICM when stackcalls are present.