(NFC) Speed up vISA asserts
Early exit in vISA assert if the condition is true so that we don't pay the
overhead of processing the assert message. Also disable vISA asserts in
release build of offline executable; it's still enabled for ReleaseInternal
build. The early exit reduces the compilation time for a large kernel from
17min to 10min for ReleaseInternal build, while disabling asserts in Release
build reduces compile time by more than 80% for the same kernel.