SelectiveHashOptions flag bug
If IGC_GET_FLAG_VALUE() is called on a flag set in SelectiveHashOptions
before g_CurrentShaderHash is set in igc_regkeys.cpp, the default
flag value is returned instead of the intended value. This solution sets
g_CurrentShaderHash as soon as the hash is calculated and if some flag
still runs into the issue - emits a warning to OptionsLog file.