Corrections to register offset calculation for payload construction
When compiling a 3D shader with SIMD width lower (for example SIMD8) than the native SIMD
execution width (for example SIMD16), register offset calculation is always 0
as it is calculated as SIMD width/native SIMD execution width. As a result, multiple mov
instructions required for setting up the payload overwrite
the same register, which is incorrect. This PR corrects the logic
by ensuring that the register offset calcluation returns at least 1.