IGA: Fixed call/calla dst footprint for SWSB
Add a trick to correct the call dst footprint that its dst type is implicit dw
but it actually writes to continuous two dw. This causes the issue
for SIMD1 case so double the exec_size to correct the case. For other simd
width, it's ok since we already conservative assume the dst footprint by
times of exec_size, which is not neccessary for call/calla that their dst
is scalar 2 dw.