Add handling for WaveShuffleIndex and correct ExtractElem/InsertElem implementation for DP emu
This change add handling for `WaveShuffleIndex` intrinsic
that change `GenISA.WaveShuffleIndex.f64` to `GenISA.WaveShuffleIndex.i64`.
Without this conversion, vISA creates mov with i64 type
that is not supported by platforms that cannot i64 support.
We change `double` -> `i64` when we don't use math operations.

In addition, this change also simplifies the implementation of
`ExtractElement` and `InsertElement`.