Flag for calculating dump hash based on kernel names
Current mechanism in IGC ShaderDump calculates the hash of the input module based on the contents of the input module - SPIR-V or OpenCL-C.

It is a common debug case to play with the original kernel content or change the higher-level compiler version like with DPCPP. In these cases the computed hash of the module will change, which in case of multiple compilation makes it hard to map the original module to the changed one.

This flag us meant to resolve this issue by changing the hash
calculation method to be based purely on entry points' names.