Reduce redundant calls to add SIMD interference with callee's maydef set
In current implementation, we gather all active default and non-default
live-ranges at each call site and mark them in a bitset stored per
FuncInfo*. After marking the bitset for all call sites, we iterate over
each FuncInfo* and mark interference between each bit set from the
bitset and maydef set of FuncInfo.

With this change, we check whether an active interval (default or
non-default) is already live-in and live-out of FuncInfo*. We don't set
the bit if it's already live-in and live-out because it means that
traditional interference has already marked edges needed and no extra
work is needed. This saves us from marking redundant edges with bits of
maydef set for callee.