Fix the DST RB calculation of Intrinsics.
We started to use intrinsic to reserve registers for some WA
implementations like barrier WAs and IEEE exception trap. However, the
dst of the intrinsic does not have the correct right bound calculated.
This might create some potential problems if the unreferenced part is
recycled by RA for other uses.