Avoid unnecessary legalization
Store with element size 64/32/16/8 bits are always good. Previous code
split 64bit into two 2xi32, which is unecessary (introducing more mov
instructions).
     store i64 ...
 to
     store <2xi32> ...
The code was probably okay for ALU instructions, but the logic does not
apply to store instructions.

The fix is to keep 64bit without splitting (same for 32/16/8bits).